requirements.txt
beautifulsoup4>=4.13.4
email-validator>=2.2.0
flask>=3.1.1
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
requests>=2.32.4
trafilatura>=2.0.0
werkzeug>=3.1.3
pymysql>=1.1.0
python-dotenv>=1.0.0
pillow>=10.0.0